+++ History.md
@@ -8,6 +8,7 @@
 - If amendments are set up to be restricted to one paragraph, then each paragraph in the motion now shows a direct link to the amendment creation page, with that paragraph pre-selected.
 - It can be set up so that external links and PDF-links will be opened in new, blank browser tabs/windows.
 - Bugfix: If a motion section was removed from a motion type that already had motions, then amendments to this motion could not be merged individually anymore.
+- Bugfix: If a voting block was deleted with motions/amendments assigned to it, then these motions/amendments could not be assigned to new voting blocks anymore.
 
 ## Version 4.13.0 (2023-10-29)